+++ storage/paths/local.go
@@ -36,7 +36,7 @@ type LocalStorage interface {
 
 const MetaFile = "sectorstore.json"
 
-const MinFreeStoragePercentage = 0
+const MinFreeStoragePercentage = float64(0)
 
 type Local struct {
 	localStorage LocalStorage
@@ -462,7 +462,7 @@ func (st *Local) reportStorage(ctx context.Context) {
 	}
 }
 
-func (st *Local) Reserve(ctx context.Context, sid storiface.SectorRef, ft storiface.SectorFileType, storageIDs storiface.SectorPaths, overheadTab map[storiface.SectorFileType]int, minFreePercentage int) (userRelease func(), err error) {
+func (st *Local) Reserve(ctx context.Context, sid storiface.SectorRef, ft storiface.SectorFileType, storageIDs storiface.SectorPaths, overheadTab map[storiface.SectorFileType]int, minFreePercentage float64) (userRelease func(), err error) {
 	ssize, err := sid.ProofType.SectorSize()
 	if err != nil {
 		return nil, err
@@ -522,12 +522,16 @@ func (st *Local) Reserve(ctx context.Context, sid storiface.SectorRef, ft storif
 			resvOnDisk = overhead
 		}
 
-		if stat.Available < overhead-resvOnDisk {
+		overheadOnDisk := overhead - resvOnDisk
+
+		if stat.Available < overheadOnDisk {
 			return nil, storiface.Err(storiface.ErrTempAllocateSpace, xerrors.Errorf("can't reserve %d bytes in '%s' (id:%s), only %d available", overhead, p.local, id, stat.Available))
 		}
 
-		if float64((stat.Available-(p.reserved+overhead))/stat.Available) < float64(minFreePercentage/100) {
-			return nil, storiface.Err(storiface.ErrTempAllocateSpace, xerrors.Errorf("can't reserve %d bytes in '%s' (id:%s), minmum free threshold reached", overhead, p.local, id))
+		freePercentag := (float64(stat.Available-overheadOnDisk) / float64(stat.Available)) * 100.0
+
+		if freePercentag < minFreePercentage {
+			return nil, storiface.Err(storiface.ErrTempAllocateSpace, xerrors.Errorf("can't reserve %d bytes in '%s' (id:%s), free disk percentage %f will be lower than minimum %f", overhead, p.local, id, freePercentag, minFreePercentage))
 		}
 
 		resID := sectorFile{sid.ID, fileType}
